[FF-3] - Base line setup for ustilizing react and layouts
