+++ src/components/Routes/Home.jsx
@@ -1,0 +1,3 @@
+function Home() {
+    
+}
